# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResponse {
  token: string;
  url: string;
  expiresAt: string;
}

export interface BatchUploadResponse {
  uploads: UploadResponse[];
}

export interface UploadInfo {
  token: string;
  originalName: string;
  mimeType: string;
  size: number;
  expiresAt: string;
  url: string;
}

export interface ErrorResponse {
  error: string;
}

export type UploadFileBody = {
  file: Blob;
};

export type BatchUploadFilesBody = {
  files: Blob[];
};