# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: uploads
    description: File upload operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /uploads:
    post:
      operationId: uploadFile
      tags: [uploads]
      summary: Upload a single file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/batch:
    post:
      operationId: batchUploadFiles
      tags: [uploads]
      summary: Upload multiple files (up to 5 images + 3 videos)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - files
      responses:
        "201":
          description: Files uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchUploadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{token}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get upload info
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{token}/file:
    get:
      operationId: serveUpload
      tags: [uploads]
      summary: Serve the uploaded file
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The file content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UploadResponse:
      type: object
      properties:
        token:
          type: string
        url:
          type: string
        expiresAt:
          type: string
      required:
        - token
        - url
        - expiresAt
    BatchUploadResponse:
      type: object
      properties:
        uploads:
          type: array
          items:
            $ref: "#/components/schemas/UploadResponse"
      required:
        - uploads
    UploadInfo:
      type: object
      properties:
        token:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        expiresAt:
          type: string
        url:
          type: string
      required:
        - token
        - originalName
        - mimeType
        - size
        - expiresAt
        - url
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error